requirements.txt
# ── PDF Processing ──────────────────────────────────
PyMuPDF==1.24.14
pdfplumber==0.11.4

# ── AI Services ─────────────────────────────────────
openai==1.86.0

# ── Video & Image Processing ────────────────────────
moviepy==2.1.2
Pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==2.1.3
scipy==1.14.1

# ── Audio Processing ────────────────────────────────
pydub==0.25.1

# ── GPU-Accelerated Encoding ────────────────────────
ffmpeg-python==0.2.0
imageio[ffmpeg]==2.36.1

# ── Web Framework (FastAPI) ─────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.18

# ── Database ────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# ── Authentication ──────────────────────────────────
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
email-validator==2.2.0

# ── Configuration ───────────────────────────────────
pydantic==2.10.4
pydantic-settings==2.7.1

# ── Legacy Web UI (Streamlit — will be retired) ─────
streamlit==1.41.1

# ── Utilities ───────────────────────────────────────
python-dotenv==1.0.1
tqdm==4.67.1
aiohttp==3.11.11
rich==13.9.4
